refactor: split insights services by card
