+++ backend/app/services/insights/__init__.py
@@ -1,0 +1,11 @@
+"""Public service entrypoints for insights card endpoints."""
+
+from app.services.insights.income_expense_breakdown import get_income_expense_breakdown
+from app.services.insights.income_expense_flow import get_income_expense_flow
+from app.services.insights.period_glance import get_period_glance
+
+__all__ = [
+    "get_income_expense_breakdown",
+    "get_income_expense_flow",
+    "get_period_glance",
+]
